ml/requirements.txt
# Temperature Forecasting Pipeline - dependencies

pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
xgboost>=1.7.0
optuna>=3.0.0
matplotlib>=3.5.0
seaborn>=0.12.0
joblib>=1.1.0
mlflow>=2.0.0
pyyaml>=6.0
